created new fields in account for sms parsing minor design changes

+++ CrystalMoneyProject/CrystalMoney/src/main/java/ru/nordmine/crystalmoney/db/MyDb.java
@@ -7,7 +7,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class MyDb extends SQLiteOpenHelper {
 
 	private static final String DATABASE_NAME = "ru.nordmine.crystalmoney.db";
-	private static final int DATABASE_VERSION = 15;
+	private static final int DATABASE_VERSION = 16;
 
 	public static final String UID = "_id";
 	
@@ -17,6 +17,8 @@ public class MyDb extends SQLiteOpenHelper {
 	public static final String ACCOUNT_COMMENT = "comment";
 	public static final String ACCOUNT_IS_CARD = "is_card";
 	public static final String ACCOUNT_PICTURE = "picture_id";
+    public static final String ACCOUNT_SMS_SENDER = "sms_sender";
+    public static final String ACCOUNT_CARD_NUMBER = "card_number";
 	
 	public static final String TRX_TABLE_NAME = "transactions";
 	public static final String TRX_TYPE = "trx_type";
@@ -43,7 +45,9 @@ public class MyDb extends SQLiteOpenHelper {
 			+ ACCOUNT_AMOUNT + " numeric not null, "
 			+ ACCOUNT_COMMENT + " varchar(255) not null, "
 			+ ACCOUNT_IS_CARD + " integer not null, "
-			+ ACCOUNT_PICTURE + " integer not null)";
+			+ ACCOUNT_PICTURE + " integer not null, "
+            + ACCOUNT_CARD_NUMBER + " varchar(50) not null, "
+            + ACCOUNT_SMS_SENDER + " varchar(50) not null)";
 	
 	private static final String SQL_CREATE_TRX = 
 			"create table " + TRX_TABLE_NAME
@@ -97,14 +101,19 @@ public class MyDb extends SQLiteOpenHelper {
         db.execSQL(SQL_CREATE_EXCHANGE);
 	}
 
-	@Override
-	public void onUpgrade(SQLiteDatabase db, int arg1, int arg2) {
-        String[] tablesForDelete = new String[]{TRX_TABLE_NAME, EXCHANGE_TABLE_NAME,
-                ACCOUNT_TABLE_NAME, CAT_TABLE_NAME};
-        for (String tableName : tablesForDelete) {
-			db.execSQL("drop table if exists " + tableName);
-		}
-		onCreate(db);
-	}
+    @Override
+    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        if (oldVersion == 15) {
+            db.execSQL("alter table " + ACCOUNT_TABLE_NAME + " add column " + ACCOUNT_CARD_NUMBER + " varchar(50) default '' not null");
+            db.execSQL("alter table " + ACCOUNT_TABLE_NAME + " add column " + ACCOUNT_SMS_SENDER + " varchar(50) default '' not null");
+        } else {
+            String[] tablesForDelete = new String[]{TRX_TABLE_NAME, EXCHANGE_TABLE_NAME,
+                    ACCOUNT_TABLE_NAME, CAT_TABLE_NAME};
+            for (String tableName : tablesForDelete) {
+                db.execSQL("drop table if exists " + tableName);
+            }
+            onCreate(db);
+        }
+    }
 
 }
